requirements (1).txt
# Core Framework
gradio>=5.49.1

# Neo4j
neo4j>=5.14.0

# AI/ML Models
transformers>=4.35.0
torch>=2.0.0
sentence-transformers>=2.2.2
openai-whisper>=20231117
openai>=1.3.0
rank_bm25>=0.2.2
scikit-learn>=1.4.2
nltk>=3.8.1
tabulate>=0.9.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.25.1
soundfile>=0.12.0
librosa>=0.10.0
# TTS>=0.22.0

# Hugging Face
huggingface_hub>=0.19.0

# NLP & Translation
langdetect>=1.0.9
deep-translator>=1.11.4

# Vector Database
faiss-cpu>=1.7.4
numpy>=1.24.0

# Speech
gTTS>=2.4.0
pydub>=0.25.1

# Utilities
python-dotenv>=1.0.0
pandas>=2.0
Pillow>=10.0.0

# Visualization
matplotlib>=3.7.0
plotly>=5.17.0
networkx>=3.1

# Additional
requests>=2.31.0
tqdm>=4.66.0
urllib3>=2.0.0
datasets>=2.14.0

# These provide fallbacks and better TTS quality
edge-tts>=6.1.0                 # Free Microsoft TTS - HIGHLY RECOMMENDED!
pyttsx3>=2.90                   # Offline TTS fallback

# ==========================================
# PDF PROCESSING
# ==========================================
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-docx>=1.1.0

# Session Management
PyJWT==2.8.0

#Excel
openpyxl>=3.1.2

